Summary view

See merge request root/billing-system!33

+++ src/main/java/com/ppp/billing/repository/BindingTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ppp.billing.model.BindingType;
 
 
-public interface BindingTypeRepository extends JpaRepository<BindingType, Long> {
+public interface BindingTypeRepository extends JpaRepository<BindingType, Integer> {
 
 	Optional<BindingType>  findByName(String name);
 	Optional<BindingType>  findById(int id);
